+++ gm-conduit/gasnet_extended_firehose.c
@@ -1,5 +1,5 @@
-/* $Id: gasnet_extended_firehose.c,v 1.29 2004/01/05 16:20:01 bonachea Exp $
- * $Date: 2004/01/05 16:20:01 $
+/* $Id: gasnet_extended_firehose.c,v 1.30 2004/01/19 11:13:58 bonachea Exp $
+ * $Date: 2004/01/19 11:13:58 $
  * Description: GASNet GM conduit Firehose DMA Registration Algorithm
  * Copyright 2002, Christian Bell <csbell@cs.berkeley.edu>
  * Terms of use are as specified in license.txt
@@ -113,7 +113,7 @@ gasnete_fh_callback_put(struct gm_port *p, void *context,
 	GASNETI_TRACE_PRINTF(C, 
 	    ("Firehose decrement remote refcount for (%p,%d) on node %d (op=%p,%p,%d)\n",
 	     (void *) pop->dest, pop->len, (unsigned) pop->req_remote.node, 
-	     (void *) pop, (void *)pop->req_remote.addr, pop->req_remote.len));
+	     (void *) pop, (void *)pop->req_remote.addr, (int)pop->req_remote.len));
 
 	fhreqs[0] = &(pop->req_remote);
 
@@ -218,7 +218,7 @@ gasnete_put_nb_bulk (gasnet_node_t node, void *dest, void *src,
 	gasnet_handle_t	handle;
 	GASNETI_TRACE_PRINTF(C, 
 	    ("gasnete_put_nb_bulk Firehose (%d,%p <- %p,%d bytes)",
-	    (unsigned) node, dest, src, nbytes));
+	    (unsigned) node, dest, src, (int)nbytes));
 
 	handle = gasnete_firehose_put_bulk(node, dest, src, nbytes, 
 		    NULL GASNETE_THREAD_PASS);
@@ -234,7 +234,7 @@ gasnete_put_nbi_bulk (gasnet_node_t node, void *dest, void *src,
 
 	GASNETI_TRACE_PRINTF(C, 
 	    ("gasnete_put_nbi_bulk Firehose (%d,%p <- %p,%d bytes)",
-	    (unsigned) node, dest, src, nbytes));
+	    (unsigned) node, dest, src, (int)nbytes));
 
 	gasnete_firehose_put_bulk(node, dest, src, nbytes, iop GASNETE_THREAD_PASS);
 	return;
@@ -293,7 +293,7 @@ gasnete_put_nb (gasnet_node_t node, void *dest, void *src,
 
 	GASNETI_TRACE_PRINTF(C, 
 	    ("gasnete_put_nb Firehose (%d,%p <- %p,%d bytes)",
-	    (unsigned) node, dest, src, nbytes));
+	    (unsigned) node, dest, src, (int)nbytes));
 
 	handle = gasnete_firehose_put(node, dest, src, nbytes, 
 		    NULL GASNETE_THREAD_PASS);
@@ -308,7 +308,7 @@ gasnete_put_nbi(gasnet_node_t node, void *dest, void *src,
 	gasnete_iop_t *iop = mythread->current_iop;
 	GASNETI_TRACE_PRINTF(C, 
 	    ("gasnete_put_nbi Firehose (%d,%p <- %p,%d bytes)",
-	    (unsigned) node, dest, src, nbytes));
+	    (unsigned) node, dest, src, (int)nbytes));
 
 	gasnete_firehose_put(node, dest, src, nbytes, iop GASNETE_THREAD_PASS);
 
@@ -379,7 +379,7 @@ extern int firehose_remote_callback(gasnet_node_t node,
 	GASNETI_TRACE_PRINTF(C, 
 	    ("Firehose RDMA PUT(rev) %p <- (%d,%p) (%d bytes)", 
 	     (void *) args->local_addr, node, (void *) args->remote_addr, 
-	     args->nbytes));
+	     (int)args->nbytes));
 
 	GASNETC_GM_PUT(_gmc.port, (void *) args->remote_addr,
 	   (gm_remote_ptr_t) args->local_addr, (unsigned long) args->nbytes,
@@ -576,7 +576,7 @@ gasnete_get_nb_bulk (void *dest, gasnet_node_t node, void *src,
 
 	GASNETI_TRACE_PRINTF(C, 
 	    ("gasnete_get_nb_bulk Firehose (%d,%p <- %p,%d bytes)",
-	    (unsigned) node, dest, src, nbytes));
+	    (unsigned) node, dest, src, (int)nbytes));
 
 	return gasnete_firehose_get(dest, node, src, nbytes, 
 		    NULL GASNETE_THREAD_PASS);
@@ -593,7 +593,7 @@ gasnete_get_nbi_bulk (void *dest, gasnet_node_t node, void *src,
 
 	GASNETI_TRACE_PRINTF(C, 
 	    ("gasnete_get_nbi_bulk Firehose (%d,%p <- %p,%d bytes)",
-	    (unsigned) node, dest, src, nbytes));
+	    (unsigned) node, dest, src, (int)nbytes));
 	gasnete_firehose_get(dest, node, src, nbytes, iop GASNETE_THREAD_PASS);
 
 	return;
